+++ sdk/api/v1alpha1/types.go
@@ -7,10 +7,14 @@ import (
 const (
 	// ProviderAws represents the new AWS provider which is implemented with Cluster API.
 	ProviderAws Provider = "aws"
+
+	// ProviderAzure represents the new Azure provider which is implemented with Cluster API.
+	ProviderAzure Provider = "azure"
 )
 
 var SupportedProviders = []Provider{
 	ProviderAws,
+	ProviderAzure,
 }
 
 type Provider string