+++ sdk/CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add Azure provider to the list of supported providers.
+
+## [0.4.0] - 2024-06-21
+
+### Added
+
 - `Client`
   - `GetReleasesForGitReference` returns all releases for the specified provider and from the specified git reference.
   - `GetReleaseForGitReference` returns a release with the specified release version for the specified provider and from